♻️ refactor documetation page related files
